+++ iOS/TimeMachineLetter/View/ClosedViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ClosedViewController.swift
+//  TimeMachineLetter
+//
+//  Created by Dean on 2022/05/24.
+//
+
+import UIKit
+
+class ClosedViewController: UIViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .red
+        
+        NSLayoutConstraint.activate([
+            
+        ])
+    }
+}
